+++ crates/hyperswitch_domain_models/Cargo.toml
@@ -19,6 +19,10 @@ common_enums = { version = "0.1.0", path = "../common_enums" }
 common_utils = { version = "0.1.0", path = "../common_utils" }
 masking = { version = "0.1.0", path = "../masking" }
 diesel_models = { version = "0.1.0", path = "../diesel_models", features = ["kv_store"] }
+cards = {version = "0.1.0", path = "../cards"}
+router_derive = {version = "0.1.0", path = "../router_derive"}
+mime = "0.3.17"
+url = { version = "2.5.0", features = ["serde"] }
 
 # Third party deps
 async-trait = "0.1.79"
@@ -27,4 +31,5 @@ http = "0.2.12"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.115"
 thiserror = "1.0.58"
+actix-web = "4.5.1"
 time = { version = "0.3.35", features = ["serde", "serde-well-known", "std"] }
