+++ crates/router/src/connector/worldpay.rs
@@ -24,6 +24,7 @@ use crate::{
     types::{
         self,
         api::{self, ConnectorCommon, ConnectorCommonExt},
+        transformers::ForeignTryFrom,
         ErrorResponse, Response,
     },
     utils::BytesExt,
@@ -225,7 +226,7 @@ impl ConnectorIntegration<api::Void, types::PaymentsCancelData, types::PaymentsR
                 Ok(types::PaymentsCancelRouterData {
                     status: enums::AttemptStatus::Voided,
                     response: Ok(types::PaymentsResponseData::TransactionResponse {
-                        resource_id: types::ResponseId::try_from(response.links)?,
+                        resource_id: types::ResponseId::foreign_try_from(response.links)?,
                         redirection_data: None,
                         mandate_reference: None,
                         connector_metadata: None,
@@ -392,7 +393,7 @@ impl ConnectorIntegration<api::Capture, types::PaymentsCaptureData, types::Payme
                 Ok(types::PaymentsCaptureRouterData {
                     status: enums::AttemptStatus::Charged,
                     response: Ok(types::PaymentsResponseData::TransactionResponse {
-                        resource_id: types::ResponseId::try_from(response.links)?,
+                        resource_id: types::ResponseId::foreign_try_from(response.links)?,
                         redirection_data: None,
                         mandate_reference: None,
                         connector_metadata: None,
