+++ crates/router/src/core/payments/flows/reject_flow.rs
@@ -3,7 +3,7 @@ use async_trait::async_trait;
 use super::{ConstructFlowSpecificData, Feature};
 use crate::{
     core::{
-        errors::{api_error_response::NotImplementedMessage, ApiErrorResponse, RouterResult},
+        errors::{ApiErrorResponse, NotImplementedMessage, RouterResult},
         payments::{self, access_token, helpers, transformers, PaymentData},
     },
     routes::AppState,
